refactor: useCardNumber 내 특정 코드 유틸로 분리
